requirements.txt
-i https://pypi.org/simple
alembic==1.0.0
beautifulsoup4==4.6.0
boto3==1.7.64
botocore==1.10.64
certifi==2018.4.16
chardet==3.0.4
click==6.7
docutils==0.14
flask-migrate==2.2.1
flask-sqlalchemy==2.3.2
flask-wtf==0.14.2
flask==1.0.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
jinja2==2.10
jmespath==0.9.3
mako==1.0.7
markupsafe==1.0
psycopg2-binary==2.7.5
psycopg2==2.7.5
python-dateutil==2.7.3; python_version >= '2.7'
python-dotenv==0.8.2
python-editor==1.0.3
python-frontmatter==0.4.2
pyyaml==3.13
requests==2.19.1
s3transfer==0.1.13
six==1.11.0
sqlalchemy==1.2.10
urllib3==1.23
werkzeug==0.14.1
wtforms==2.2.1
